Separate StorageManager into Tablespace and specific StorageManager.
