+++ tajo-core/src/main/java/org/apache/tajo/engine/planner/physical/StoreTableExec.java
@@ -30,10 +30,7 @@ import org.apache.tajo.catalog.statistics.TableStats;
 import org.apache.tajo.plan.logical.InsertNode;
 import org.apache.tajo.plan.logical.PersistentStoreNode;
 import org.apache.tajo.plan.util.PlannerUtil;
-import org.apache.tajo.storage.Appender;
-import org.apache.tajo.storage.FileStorageManager;
-import org.apache.tajo.storage.StorageManager;
-import org.apache.tajo.storage.Tuple;
+import org.apache.tajo.storage.*;
 import org.apache.tajo.unit.StorageUnit;
 import org.apache.tajo.worker.TaskAttemptContext;
 
@@ -93,7 +90,7 @@ public class StoreTableExec extends UnaryPhysicalExec {
         lastFileName = new Path(lastFileName + "_" + suffixId);
       }
 
-      appender = ((FileStorageManager)StorageManager.getFileStorageManager(context.getConf()))
+      appender = ((FileStorageManager) TableSpaceManager.getFileStorageManager(context.getConf()))
           .getAppender(meta, appenderSchema, lastFileName);
 
       if (suffixId > 0) {
@@ -101,7 +98,7 @@ public class StoreTableExec extends UnaryPhysicalExec {
             "The remain output will be written into " + lastFileName.toString());
       }
     } else {
-      appender = StorageManager.getStorageManager(context.getConf(), meta.getStoreType()).getAppender(
+      appender = TableSpaceManager.getStorageManager(context.getConf(), meta.getStoreType()).getAppender(
           context.getQueryContext(),
           context.getTaskId(), meta, appenderSchema, context.getQueryContext().getStagingDir());
     }
