+++ tajo-core/src/main/java/org/apache/tajo/worker/TajoWorker.java
@@ -55,7 +55,7 @@ import org.apache.tajo.rule.EvaluationFailedException;
 import org.apache.tajo.rule.SelfDiagnosisRuleEngine;
 import org.apache.tajo.rule.SelfDiagnosisRuleSession;
 import org.apache.tajo.storage.HashShuffleAppenderManager;
-import org.apache.tajo.storage.StorageManager;
+import org.apache.tajo.storage.TableSpaceManager;
 import org.apache.tajo.util.*;
 import org.apache.tajo.util.history.HistoryReader;
 import org.apache.tajo.util.history.HistoryWriter;
@@ -368,7 +368,7 @@ public class TajoWorker extends CompositeService {
     }
 
     try {
-      StorageManager.shutdown();
+      TableSpaceManager.shutdown();
     } catch (IOException ie) {
       LOG.error(ie.getMessage(), ie);
     }
