+++ tajo-core/src/test/java/org/apache/tajo/engine/query/TestJoinBroadcast.java
@@ -569,7 +569,7 @@ public class TestJoinBroadcast extends QueryTestCaseBase {
         }
         Path dataPath = new Path(table.getPath().toString(), fileIndex + ".csv");
         fileIndex++;
-        appender = ((FileStorageManager)StorageManager.getFileStorageManager(conf))
+        appender = ((FileStorageManager) TableSpaceManager.getFileStorageManager(conf))
             .getAppender(tableMeta, schema, dataPath);
         appender.init();
       }
